+++ R/QIM_report_UI.R
@@ -109,7 +109,7 @@ qim_reportCreator_UI <- function(id) {
       shiny::column(
         width = 4,
         shiny::wellPanel(
-          style = "height:23em",
+          style = "height:25em",
           shiny::tags$h5("Create report"),
           shiny::fluidRow(
             shiny::column(
@@ -133,24 +133,6 @@ qim_reportCreator_UI <- function(id) {
             )
           ),
           shiny::hr(),
-          shiny::textInput(
-            inputId = ns("filename"),
-            label = "Name for file",
-            value = paste0("QIMReport-", Sys.Date(), ".csv")
-          ),
-          shinyjs::disabled(
-            # enable when there is something to save...
-            shiny::downloadButton(
-              outputId = ns("download_button"),
-              label = "Download"
-            )
-          )
-        )
-      ),
-      shiny::column(
-        width = 4,
-        shiny::wellPanel(
-          style = "height:23em",
           shiny::tags$h5("Number of reports"),
           shiny::fluidRow(
             shiny::column(
@@ -182,6 +164,39 @@ qim_reportCreator_UI <- function(id) {
             )
           )
         )
+      ),
+      shiny::column(
+        width = 4,
+        shiny::wellPanel(
+          style = "height:25em",
+          shiny::tags$h5("Report store/save"),
+          shiny::hr(),
+          shiny::tags$h6("CSV 'spreadsheet' download"),
+          shiny::icon("download"),
+          shinyjs::disabled(
+            # enable when there is something to save...
+            shinyFiles::shinySaveButton(
+              id = ns("csv_filename"),
+              label = "Download CSV file",
+              title = "Download CSV file - choose CSV filename",
+              filename = paste0("QIMReport-", Sys.Date()),
+              filetype = list(spreadsheet = c('csv'))
+            )
+          ),
+          shiny::hr(),
+          shiny::tags$h6("JSON 'PIP' download"),
+          shiny::icon("download"),
+          shinyjs::disabled(
+            # enable when there is something to save...
+            shinyFiles::shinySaveButton(
+              id = ns("json_filename"),
+              label = "Download JSON file",
+              title = "Download JSON file - choose JSON filename",
+              filename = paste0("QIMReport-", Sys.Date()),
+              filetype = list(json = c('json'))
+            )
+          )
+        )
       )
     )
   )
@@ -194,6 +209,8 @@ qim_reportCreator_UI <- function(id) {
 #' @param session as required by Shiny modules
 #' @param dMQIM dMeasure QIM R6 object
 #'  access to appointments lists, results, correspondence and EMR database
+#' @param report a list returned by qim_reportCharter
+#'   should contain $report_values(), which is a dataframe
 #'
 #' @return list with following components
 #' \describe{
@@ -201,8 +218,18 @@ qim_reportCreator_UI <- function(id) {
 #' }
 #'
 #' @export
-qim_reportCreator <- function(input, output, session, dMQIM) {
+qim_reportCreator <- function(input, output, session, dMQIM, report) {
   ns <- session$ns
+
+  volumes <- c(Home = fs::path_home(), shinyFiles::getVolumes()())
+  shinyFiles::shinyFileSave(
+    input, id = "csv_filename", roots = volumes, session = session,
+    restrictions = system.file(package = "base")
+  )
+  shinyFiles::shinyFileSave(
+    input, id = "json_filename", roots = volumes, session = session,
+    restrictions = system.file(package = "base")
+  )
 
   empty_result <- data.frame(
     QIM = character(),
@@ -223,9 +250,11 @@ qim_reportCreator <- function(input, output, session, dMQIM) {
     ignoreInit = FALSE, ignoreNULL = FALSE, {
       if (is.null(report_values()) || nrow(report_values()) == 0) {
         # disable download button if nothing to download
-        shinyjs::disable("download_button")
+        shinyjs::disable("csv_filename")
+        shinyjs::disable("json_filename")
       } else {
-        shinyjs::enable("download_button")
+        shinyjs::enable("csv_filename")
+        shinyjs::enable("json_filename")
       }
     }
   )
@@ -483,12 +512,86 @@ qim_reportCreator <- function(input, output, session, dMQIM) {
       }
     })
 
-  output$download_button <- shiny::downloadHandler(
-    filename = function() {
-      input$filename
-    },
-    content = function(file) {
-      write.csv(report_values(), file = file, row.names = FALSE)
+  shiny::observeEvent(
+    input$csv_filename,
+    ignoreInit = TRUE, ignoreNULL = TRUE, {
+      shiny::req(input$csv_filename) # can't be NULL/empty
+
+      datapath <- shinyFiles::parseSavePath(volumes, input$csv_filename) %>>%
+        dplyr::pull(datapath)
+      if (length(datapath)) {
+        # if length 0 (i.e. datapath == character(0), then return empty string)
+        write.csv(report_values(), file = datapath, row.names = FALSE)
+      }
+    }
+  )
+
+  # if json_filename clicked,
+  # then ask various questions about which data to export
+  # and author/practice ID and small cell suppression
+  shiny::observeEvent(
+    input$json_filename,
+    ignoreInit = TRUE, ignoreNULL = TRUE, {
+      shiny::req(input$json_filename) # can't be NULL/empty
+      datapath <- shinyFiles::parseSavePath(volumes, input$json_filename) %>>%
+        dplyr::pull(datapath)
+
+      unique_DateTo <- unique(report_values()$DateTo)
+
+      if (length(datapath)) {
+        # if length 0 (i.e. datapath == character(0), then return empty string)
+
+        shiny::showModal(shiny::modalDialog(
+          title = "Practice Incentive Program JSON attributes",
+          shiny::selectInput(
+            inputId = ns("json_DateTo"),
+            label = "'Date To' of period",
+            choices = unique_DateTo
+          ),
+          shiny::textInput(
+            inputId = ns("json_author_id"),
+            label = "Author ID",
+            value = "bpsrawdata"
+          ),
+          shiny::textInput(
+            inputId = ns("json_practice_id"),
+            label = "Practice ID",
+            value = "",
+            placeholder = "Your practice ID"
+          ),
+          shiny::checkboxInput(
+            inputId = ns("json_small_cell_suppression"),
+            label = "Small cell suppression",
+            value = TRUE
+          ),
+          easyClose = FALSE,
+          footer = shiny::tagList(
+            shiny::modalButton("Cancel"),
+            shiny::actionButton(inputId = ns("json_ok"), "Save JSON")
+          )
+        ))
+      }
+    })
+
+  # if 'Save JSON' button clicked in write JSON modal dialog
+  # then create JSON string and output to file
+  shiny::observeEvent(
+    input$json_ok,
+    ignoreInit = TRUE, ignoreNULL = TRUE, {
+
+      datapath <- shinyFiles::parseSavePath(volumes, input$json_filename) %>>%
+        dplyr::pull(datapath)
+
+      json_string <- writeReportJSON(
+        d = report_values(),
+        date_to = input$json_DateTo,
+        author_id = input$json_author_id,
+        practice_id = input$json_practice_id,
+        small_cell_suppression = input$json_small_cell_suppression,
+      )
+      write(json_string, file = datapath)
+
+      shiny::removeModal()
     }
   )
 
@@ -586,6 +689,28 @@ qim_reportCreator <- function(input, output, session, dMQIM) {
           timeOut = 10000
         )
         n_report_warning(TRUE)
+      }
+    }
+  )
+
+  # if report
+  shiny::observeEvent(
+    report$report_values(),
+    ignoreInit = TRUE, ignoreNULL = TRUE, {
+      shiny::req(report$report_values())
+
+      if (nrow(report$report_values()) > 0) {
+        report_values(
+          report$report_values() %>>%
+            dplyr::mutate(DateTo = as.character(DateTo))
+          # co-erce to character (instead of numeric)
+        )
+        # copy the dataframe
+        shinytoastr::toastr_info(
+          "Copying report to Report Creator, can be saved as JSON",
+          closeButton = TRUE,
+          position = "bottom-left", title = "PIP report"
+        )
       }
     }
   )